fastapi-assessment/requirements.txt
fastapi
uvicorn
sqlalchemy
mediapipe
pydantic
python-jose
passlib
